requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
celery
redis
python-jose[cryptography]
passlib[bcrypt]
transformers
torch
Pillow
insightface
onnxruntime
opencv-python
requests
razorpay
ffmpeg-python
numpy
supabase
